parser code organisation
