+++ src/parser/parser.py
@@ -1,18 +1,9 @@
 from textx import metamodel_from_file
 import os
+from src.parser import init_cv_metamodel
 
 
-def basic_info_processor(basic_info):
-    if basic_info.picture and not os.path.exists(basic_info.picture):
-        raise ValueError(f"Picture file does not exist: {basic_info.picture}")
-
-
-cv_grammar_path = 'src/dsl/cv.tx'
-cv_metamodel = metamodel_from_file(cv_grammar_path)
-
-cv_metamodel.register_obj_processors({
-    'BasicInfo': basic_info_processor
-})
+cv_metamodel = init_cv_metamodel()
 
 
 def parse_cv(file_path):
